Update : add printf, gnl and dll

+++ libft/src/dll_insert_head.c
@@ -3,19 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   dll_insert_head.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lagea < lagea@student.s19.be >             +#+  +:+       +#+        */
+/*   By: lagea <lagea@student.s19.be>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/02 13:18:21 by lagea             #+#    #+#             */
-/*   Updated: 2024/05/21 22:27:26 by lagea            ###   ########.fr       */
+/*   Updated: 2024/05/20 18:33:44 by lagea            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/libft.h"
+#include "../../inc/push_swap.h"
 
-void dll_insert_head(int data, t_stack *stack)
+void	dll_insert_head(ssize_t data, t_stack *stack)
 {
-	t_node *new;
-	
+	t_node	*new;
+
 	new = dll_new_node(data);
 	if (stack->head == NULL)
 	{
@@ -26,6 +27,6 @@ void dll_insert_head(int data, t_stack *stack)
 	{
 		new->next = stack->head;
 		stack->head->prev = new;
-		stack->head = new;	
+		stack->head = new;
 	}
 }
